fixed some bugs and added some all useful programs

+++ configure.ps1
@@ -7,6 +7,7 @@ $tweaks = @(
     "InstallJava",
 	"InstallWSL2",
 	"InstallUbuntuForWSL",
+    "InstallGit",
 	"InstallPutty",
 	"InstallWinSCP",
 	"InstallNeovim",
@@ -14,7 +15,8 @@ $tweaks = @(
 	"InstallThunderbird",
 	"InstallLinkShellExtension",
 	"InstallUninstallView",
-	"InstallQBitTorrent"
+	"InstallQBitTorrent",
+    "FinishedInstall"
 )
 
 $global:IsInitialized = 0
@@ -24,40 +26,64 @@ $global:DriveLetters
 
 # installing Chocolatey and provide the option to move the default install drive
 Function InstallChoco {
-	$Title = ""
-	$Message = "To Install Chocolatey hit I or R to also change the installation drive for all chocolatey installed programs"
-	$Options = "&Install", "&RelocateAndInstall"
+$Title = ""
+	$Message = "To Install Chocolatey hit I or R to also change the installation drive"
+	$Options = "&Install", "&RelocateAndInstall", "&Skip"
 	Write-Output "About to install Chocolatey"
-    Write-Warning -Message "This is required for installing programs with this script"
-	$DefaultChoice = 0
+    Write-Warning -Message "This is required for base programs"
+	$DefaultChoice = 1
 	$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
+    $ChocoPath=":\Program Files (x86)\Chocolatey"
+    $ChocoEnv="ChocolateyInstall"
+    $ChocoToolsEnv="ChocolateyToolsLocation"
+    $DoChocoInstall=1
+    $IsChocoInPath=Test-Path -Path "$env:ChocolateyInstall" -ErrorAction SilentlyContinue
 
-	if($Result -eq 1){
-		initDrives
+	if($Result -le 1){
+		if($Result -eq 1){
+        initDrives
 		$Title = "`nSelect the drive where Chocolatey should be installed"
 		$SelectedDrive = ShowMenu -Title $Title -Menu $global:DriveLetters -Default $global:Default
-		$New_Location = "${SelectedDrive}:\Program Files (x86)\Chocolatey"
-		if ("$env:ChocolateyInstall" -AND $SelectedDrive -ne $env:ChocolateyInstall.Substring(0,1)) {
+		if ("$env:ChocolateyInstall" -AND $SelectedDrive -ne $env:ChocolateyInstall.Substring(0,1) -and !(Test-Path -Path "${SelectedDrive}$ChocoPath" -ErrorAction SilentlyContinue) -and $IsChocoInPath) {
+            Write-Output "Chocolatey is installed on another drive..."
+            Write-Output "Moving Chocolatey installation to selected drive..."
 			Move-Item -force "$env:ChocolateyInstall" "${SelectedDrive}:\Program Files (x86)"
-		} elseif (!("$env:ChocolateyInstall")) {
-			$Result = 0
-		}
-		[System.Environment]::SetEnvironmentVariable("ChocolateyInstall","${SelectedDrive}:\Program Files (x86)\Chocolatey",[System.EnvironmentVariableTarget]::Machine)
-		Write-Output "Parsing Environment Variable..."
-		# reinit environment variable
-		$env:ChocolateyInstall = [System.Environment]::GetEnvironmentVariable("ChocolateyInstall","Machine")	
-		}
-	if($Result -eq 0){
-		if (!($env:ChocolateyInstall) -or !(Test-Path -Path "$env:ChocolateyInstall" -ErrorAction SilentlyContinue)){
-			Write-Output "Installing Chocolatey..."
-			Set-ExecutionPolicy Bypass -Scope Process -Force; [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072; iex ((New-Object System.Net.WebClient).DownloadString('https://chocolatey.org/install.ps1'))
-			choco install chocolatey-core.extension -y
-			} Else {
-				Write-Output "Chocolatey is already installed, skipping..."
-			}
-		}
+            $DoChocoInstall=0
+            }
+         }
+            if(!($SelectedDrive)){
+                $SelectedDrive=((Get-Location).Path.Substring(0,1))
+            }
+        
+		    Write-Output "Parsing Environment Variable..."
+            [System.Environment]::SetEnvironmentVariable("$ChocoToolsEnv", "${SelectedDrive}$ChocoPath\tools" ,[System.EnvironmentVariableTarget]::Machine)
+            [System.Environment]::SetEnvironmentVariable("$ChocoEnv","${SelectedDrive}$ChocoPath",[System.EnvironmentVariableTarget]::Machine)
+            # reinit environment variable
+		    $env:ChocolateyInstall = [System.Environment]::GetEnvironmentVariable("$ChocoEnv","Machine")
+            $env:ChocolateyToolsLocation = [System.Environment]::GetEnvironmentVariable("$ChocoToolsEnv","Machine")
+                
+            if(Test-Path -Path "${SelectedDrive}$ChocoPath"){
+                Write-Output "Chocolatey is already installed, skipping installation..."
+                $DoChocoInstall=0
+            }
+
+            if($DoChocoInstall -eq 1){
+                Write-Output "Installing Chocolatey..."
+			    Set-ExecutionPolicy Bypass -Scope Process -Force; [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072; iex ((New-Object System.Net.WebClient).DownloadString('https://chocolatey.org/install.ps1'))
+			    choco install chocolatey-core.extension -y
+            }
+		}else{
+            if($IsChocoInPath){
+                Write-Output "Chocolatey is already installed, proceeding to the next step..."
+            }else{
+                Write-Output "Chocolatey core dependency program install skipped, exiting..."
+                Start-Sleep -s 3
+                FinishedInstall
+                Exit
+            }
+        }
 	PromptInstallAll
-	}
+}
 
 # installing programs
 Function InstallJava {
@@ -89,6 +115,40 @@ Function InstallFirefox {
 		if(($Result -eq 0) -or $global:InstallEverything){
 			Write-Output "Installing Firefox..."
 			choco install firefox -y
+		}else{
+			Write-Warning -Message "Skipping..."
+		}
+}
+
+Function InstallThunderbird {
+		$Title = ""
+		$Message = "To Install Thunderbird hit I otherwise use S to skip"
+		$Options = "&Install", "&Skip"
+		$Result = 0
+		$DefaultChoice = 1
+		if (!($global:InstallEverything)){
+			$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
+		}
+		if(($Result -eq 0) -or $global:InstallEverything){
+			Write-Output "Installing Thunderbird..."
+			choco install thunderbird -y
+		}else{
+			Write-Warning -Message "Skipping..."
+		}
+}
+
+Function InstallPaintDotNet{
+		$Title = ""
+		$Message = "To Install Paint.net hit I otherwise use S to skip"
+		$Options = "&Install", "&Skip"
+		$Result = 0
+		$DefaultChoice = 1
+		if (!($global:InstallEverything)){
+			$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
+		}
+		if(($Result -eq 0) -or $global:InstallEverything){
+			Write-Output "Installing Paint.net..."
+			choco install paint.net -y
 		}else{
 			Write-Warning -Message "Skipping..."
 		}
@@ -128,6 +188,23 @@ Function InstallWSL2 {
     }
 }
 
+Function InstallNotion {
+		$Title = ""
+		$Message = "To Install Notion hit I otherwise use S to skip"
+		$Options = "&Install", "&Skip"
+		$Result = 0
+		$DefaultChoice = 1
+		if (!($global:InstallEverything)){
+			$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
+		}
+		if(($Result -eq 0) -or $global:InstallEverything){
+			Write-Output "Installing Notion..."
+			choco install notion -y
+		}else{
+			Write-Warning -Message "Skipping..."
+	}
+}
+
 Function InstallUbuntuForWSL{
 		$Title = ""
 		$Message = "To Install Ubuntu as Subsystem Distro hit I or use S to skip (This works only if you have installed WSL version 1 or 2)"
@@ -138,15 +215,16 @@ Function InstallUbuntuForWSL{
 			$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
 		}
 		if(($Result -eq 0) -or $global:InstallEverything){
-            choco install wsl-ubuntu-2004
+            choco install wsl-ubuntu-2004 -y
         }else{
 			Write-Warning -Message "Skipping..."
     }
 }
 
-Function InstallUbuntuHere{
+# install git with posh-git
+Function InstallGit{
 		$Title = ""
-		$Message = "To Install Ubuntu(WSL) Here  hit I or use S to skip (This works only if you have installed Ubuntu(WSL))"
+		$Message = "To Install Git hit I or use S to skip"
 		$Options = "&Install", "&Skip"
 		$Result = 0
 		$DefaultChoice = 1
@@ -154,7 +232,24 @@ Function InstallUbuntuHere{
 			$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
 		}
 		if(($Result -eq 0) -or $global:InstallEverything){
-            choco install ubuntuhere
+            choco install git.install --params "/GitAndUnixToolsOnPath /WindowsTerminal /NoShellIntegration /NoGuiHereIntegration /NoShellHereIntegration" -y
+            PromptInstallPoshGit
+        }else{
+			Write-Warning -Message "Skipping..."
+    }
+}
+
+Function PromptInstallPoshGit{
+		$Title = ""
+		$Message = "To Install posh-git hit I or use S to skip"
+		$Options = "&Install", "&Skip"
+		$Result = 0
+		$DefaultChoice = 1
+		if (!($global:InstallEverything)){
+			$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
+		}
+		if(($Result -eq 0) -or $global:InstallEverything){
+            choco install poshgit -y
         }else{
 			Write-Warning -Message "Skipping..."
     }
@@ -197,12 +292,12 @@ Function InstallWinSCP {
 Function InstallNeovim {
 		$Title = ""
 		$Message = "To Install Neovim stable hit I, otherwise use B to install the beta for support more keyboard layouts or use S to skip"
-		$Options = "&Install",  "BetaInstall", "&Skip"
+		$Options = "&Install",  "&BetaInstall", "&Skip"
 		$Result = 0
-		$DefaultChoice = 1
-		if (!($global:InstallEverything)){
-			$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
-		}
+		$DefaultChoice = 2
+		
+        $Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
+		
 		if($Result -eq 0){
 			Write-Output "Installing Neovim..."
 			choco install neovim -y
@@ -214,40 +309,6 @@ Function InstallNeovim {
 		}else{
 			Write-Warning -Message "Skipping..."
 	}
-}
-
-Function InstallPaintDotNet{
-		$Title = ""
-		$Message = "To Install Paint.net hit I otherwise use S to skip"
-		$Options = "&Install", "&Skip"
-		$Result = 0
-		$DefaultChoice = 1
-		if (!($global:InstallEverything)){
-			$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
-		}
-		if(($Result -eq 0) -or $global:InstallEverything){
-			Write-Output "Installing Paint.net..."
-			choco install paint.net -y
-		}else{
-			Write-Warning -Message "Skipping..."
-		}
-}
-
-Function InstallThunderbird {
-		$Title = ""
-		$Message = "To Install Thunderbird hit I otherwise use S to skip"
-		$Options = "&Install", "&Skip"
-		$Result = 0
-		$DefaultChoice = 1
-		if (!($global:InstallEverything)){
-			$Result = $Host.UI.PromptForChoice($Title, $Message, $Options, $DefaultChoice)
-		}
-		if(($Result -eq 0) -or $global:InstallEverything){
-			Write-Output "Installing Thunderbird..."
-			choco install thunderbird -y
-		}else{
-			Write-Warning -Message "Skipping..."
-		}
 }
 
 Function InstallLinkShellExtension{
@@ -321,6 +382,7 @@ Function PromptInstallAll {
 	}
 }
 
+# enable wsl2
 Function InstallLinuxSubsystemV2 {
 	if ((Get-WindowsOptionalFeature -Online -FeatureName "VirtualMachinePlatform").State -eq "Enabled"){
 		wsl --set-default-version 2
@@ -427,6 +489,10 @@ function ShowMenu
 		}
 	}
 	while ($k.Key -notin ([ConsoleKey]::Escape, [ConsoleKey]::Enter))
+}
+
+Function FinishedInstall{
+    Write-Output  "`nFinished script, you can check for errors and close the window`n"
 }
 
 # Call the desired tweak functions
